requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.7.1
yarl==1.9.2
